Moved algorithm in Event Service
